Fix signedness/uninitialized/unused issues.
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@3935 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/nbody/nbody_c/stencil_functions.hpp
@@ -14,7 +14,7 @@
 /// The function \a generate_initial_data will be called to initialize the 
 /// given instance of 'stencil_data' 
 HPX_COMPONENT_EXPORT int generate_initial_data(
-    stencil_data* data, int item, int maxitems, int row,
+    stencil_data* data, unsigned long item, int maxitems, int row,
             Par const& par);
 
 /// The function \a evaluate_timestep will be called to compute the result data
